New image base and updates to support neovim installation
For my development environment I wanted to include neovim. This made
me deviate from the openjdk image due to ease of installation.
I chose alpine since it's relatively small and then needed to update the installation
of scala and sbt accordingly. This is step one, next is config to get this fully functioning.